Implement auto update

+++ GTNHJenksinsDownloader/Program.cs
@@ -5,6 +5,8 @@ using System.Diagnostics;
 using System.IO.Compression;
 using System.Net;
 using System.Runtime.InteropServices;
+using System.Threading;
+using Newtonsoft.Json;
 
 namespace GTNHJenksinsDownloader
 {
@@ -15,26 +17,39 @@ namespace GTNHJenksinsDownloader
         /// </summary>
         /// 
 
-        [DllImport("kernel32.dll")]
-        static extern IntPtr GetConsoleWindow();
-
-        [DllImport("user32.dll")]
-        static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
-
-        const int SW_HIDE = 0;
-        const int SW_SHOW = 5;
-
         [STAThread]
-        static void Main()
+        static void Main(string[] args)
         {
-#if !DEBUG
-            var handle = GetConsoleWindow();
 
-            // Hide
-            ShowWindow(handle, SW_HIDE);
+#if DEBUG
+            Utility.CreateConsole();
+#endif
+
+            //Console.WriteLine("test");
+
+            if (Array.Exists(args, x => x == "updated"))
+            {
+                MessageBox.Show("Updated", Settings.appname, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                string temppath = Path.Combine(Path.GetTempPath(), "GTNHJenkinsDownloader\\");
+                string file = Path.Combine(temppath, "temp.exe");
+                if (File.Exists(file)) File.Delete(file);
+                if (Directory.Exists(temppath)) Directory.Delete(temppath);
+            }
+            if (Array.Exists(args, x => x == "wait"))
+                Thread.Sleep(500);
+            if (Array.Exists(args, x => x == "update"))
+            {
+                Thread.Sleep(500);
+                string file = Path.Combine(Path.GetTempPath(), "GTNHJenkinsDownloader\\temp.exe");
+                File.Copy(file, Settings.apppath, true);
+                Process.Start(Settings.apppath, "updated");
+                return;
+            }
+
+#if !DEBUG
             if (Process.GetProcessesByName(Path.GetFileNameWithoutExtension(Settings.apppath)).Length > 1)
             {
-                MessageBox.Show("App arledy running !", "GTNH Jenkins Downloader", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("App alredy running !", "GTNH Jenkins Downloader", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return;
             }
             string exepath = System.Reflection.Assembly.GetExecutingAssembly().Location;
@@ -44,13 +59,46 @@ namespace GTNHJenksinsDownloader
             {
                 if (!Directory.Exists(Settings.appdirectory))
                     Directory.CreateDirectory(Settings.appdirectory);
-                //if (!File.Exists(Settings.apppath))
-                File.Copy(exepath, Settings.apppath, true); // updates
-                Process.Start(Settings.apppath);
+                if (!File.Exists(Settings.apppath))
+                    File.Copy(exepath, Settings.apppath, true);
+                else
+                {
+                    string fver = FileVersionInfo.GetVersionInfo(Settings.apppath).ProductVersion;
+                    if (Utility.GetBuildNumberFromVersionString(fver) < Settings.buildNumber ||
+                        fver == "1.0.0.0" /* Before implementing versions */)
+                    {
+                        File.Copy(exepath, Settings.apppath, true);
+                        Process.Start(Settings.apppath, "updated");
+                        return;
+                    }
+                }
+                Process.Start(Settings.apppath, "wait");
                 return;
             }
+            else // Check for updates
+            {
+                string res = Utility.Get("https://api.github.com/repos/kuba6000/GTNH-Jenkins-Downloader/releases");
+                Newtonsoft.Json.Linq.JArray arr = (Newtonsoft.Json.Linq.JArray)JsonConvert.DeserializeObject(res);
+                if (arr.Count > 0)
+                {
+                    if (Utility.GetBuildNumberFromVersionString(((string)arr[0]["name"]).Substring(1)) > Settings.buildNumber)
+                    {
+                        DialogResult result = MessageBox.Show("Do you want to update ?\n" + Settings.version + " -> " + (string)arr[0]["name"], "New update found !", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+                        if (result == DialogResult.Yes)
+                        {
+                            string temppath = Path.Combine(Path.GetTempPath(), "GTNHJenkinsDownloader\\");
+                            if (!Directory.Exists(temppath)) Directory.CreateDirectory(temppath);
+                            string file = Path.Combine(temppath, "temp.exe");
+                            if (File.Exists(file)) File.Delete(file);
+                            WebClient client = new WebClient();
+                            client.DownloadFile((string)arr[0]["assets"][0]["browser_download_url"], file);
+                            Process.Start(file, "update");
+                            return;
+                        }
+                    }
+                }
+            }
 #endif
-            //https://github.com/JamesNK/Newtonsoft.Json/releases/download/13.0.1/Json130r1.zip
 
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
